==> python/notebooks/02_selecting_areas_from_northeast.ipynb <==
# Area selection based on Wind Spead

All automatic stations in Northeast region of Brazil will be analysed

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from geopy.distance import distance


import os

%load_ext autoreload
%autoreload 2
from data_preprocessing import initial_data_preprocessing

pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', 1000)
pd.set_option('display.width', 1000)


def extract_wind_speed_statistics(df: pd.DataFrame) -> dict:
    '''
    Extracts relevant information about the wind speed from the automatic stations.
    
    Parameters
    ----------
    df: Preprocessed dataframe containing the historic information about the wind speed and direction.
    
    Returns 
    ----------
    wind_statistics: dict containing the mean, median, std, minimum date available for the wind speed, null quantity, dataset size and non null dataset size.

    '''

    wind_statistics = {
        'WIND_SPEED_mean': df['WIND_SPEED_ms'].mean(),
        'WIND_SPEED_120m_mean': df['WIND_SPEED_120m_ms'].mean(),
        'WIND_SPEED_median': df['WIND_SPEED_ms'].median(),
        'WIND_SPEED_120m_median': df['WIND_SPEED_120m_ms'].median(),
        'WIND_SPEED_std': df['WIND_SPEED_ms'].std(),
        'WIND_SPEED_120m_std': df['WIND_SPEED_120m_ms'].std(),
        'MINIMUM_DATE': df['DATE_MEASUREMENT'].min(),
        'DATASET_SIZE': df['WIND_SPEED_ms'].shape[0],
        'NON_NULL_DATASET_SIZE': df.dropna().shape[0],
        'NON_NULL_RATIO': df.dropna().shape[0] / df['WIND_SPEED_ms'].shape[0],
        'NULL_CELLS_QUANTITY': df.isna().sum().sum(),
    }
    return wind_statistics

df_catalog_automatic_stations = pd.read_excel('./../../data/automatic_stations_catalog.xlsx')[[
                                                        'DC_NOME','SG_ESTADO','CD_SITUACAO',
                                                        'VL_LATITUDE','VL_LONGITUDE','VL_ALTITUDE',
                                                        'DT_INICIO_OPERACAO','CD_ESTACAO']]

df_catalog_automatic_stations = df_catalog_automatic_stations.query("CD_SITUACAO != 'Pane'")

northeast_stations = os.listdir('./../../data/raw')
# northeast_stations = [x.split('.')[0].replace('dados_','') for x in northeast_stations]
df_northeast_stations = pd.DataFrame(data=northeast_stations,columns=['STATION_FILE_NAME'])
df_northeast_stations['CD_ESTACAO'] = [x.split('.')[0].replace('dados_','').split('_H_')[0] for x in northeast_stations]
df_northeast_stations = df_northeast_stations.merge(df_catalog_automatic_stations,on='CD_ESTACAO')

wind_station_statistics = []
for ii,station in df_northeast_stations.iterrows():
    df = pd.read_csv("./../../data/raw/" + station['STATION_FILE_NAME'],header=9,sep=';')
    print(f"Analysing {station['CD_ESTACAO']} - {station['DC_NOME']} station \n")
    df = initial_data_preprocessing(df,minimum_date='2018-01-01',maximum_date='2021-01-01') #Filtering from 2018 to 2021 since it is the data available in Conde station
    wind_statistics = extract_wind_speed_statistics(df)
    wind_statistics['CD_ESTACAO'] = station['CD_ESTACAO']
    wind_statistics['NOME'] = station['DC_NOME']
    wind_statistics['SG_ESTADO'] = station['SG_ESTADO']
    wind_statistics['LATITUDE'] = station['VL_LATITUDE']
    wind_statistics['LONGITUDE'] = station['VL_LONGITUDE']
    wind_statistics['ALTITUDE'] = station['VL_ALTITUDE']
    wind_station_statistics.append(wind_statistics)
df_wind_statistics = pd.DataFrame(wind_station_statistics)

wind_statistics

df_wind_statistics = df_wind_statistics[['CD_ESTACAO', 'NOME', 'SG_ESTADO', 'WIND_SPEED_mean', 'WIND_SPEED_120m_mean', 'WIND_SPEED_median', 
                                         'WIND_SPEED_120m_median', 'WIND_SPEED_std', 'WIND_SPEED_120m_std', 'MINIMUM_DATE', 'DATASET_SIZE', 
                                         'NON_NULL_RATIO', 'NON_NULL_DATASET_SIZE', 'NULL_CELLS_QUANTITY', 'LATITUDE', 'LONGITUDE','ALTITUDE']]

df_wind_statistics.sort_values(by=['WIND_SPEED_median'],ascending=[False]).query("NON_NULL_RATIO >=0.75").head(10)

df_wind_statistics.to_csv("./../../data/automatic_stations_wind_statistics.csv",index=False)

# Conclusions

Station of Conde seems adequate. The city is not a capital and there is good amount of available data, since late september 2017.
