==> python/notebooks/04_joining_stations_data.ipynb <==
# Joining and Analysing data from all stations

import glob

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

%load_ext autoreload
%autoreload 2
from data_preprocessing import *

pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', 1000)
pd.set_option('display.width', 1000)

df_station_statistics = pd.read_csv("./../../data/automatic_stations_wind_statistics.csv")
station_codes = ['A431','A458', 'A409', 'A401', 'A413', 'A406', 'A442', 'A450', 'A434', 'A436']
station_names = ['CONDE','RIBEIRA DO AMPARO', 'ARACAJU', 'SALVADOR', 'FEIRA DE SANTANA', 'CRUZ DAS ALMAS', 'EUCLIDES DA CUNHA',
                 'JEREMOABO', 'AMARGOSA', 'QUEIMADAS']

df_stations = []
for station_code,station_name in zip(station_codes,station_names):
    station_file_name = glob.glob(f'./../../data/raw/*{station_code}*')[0]
    df_raw = pd.read_csv(station_file_name,header=9,sep=';')
    df = initial_data_preprocessing(df_raw,minimum_date='2018-01-01',maximum_date='2021-01-01')
    df['NAME'] = station_name
    df['CODE'] = station_code
    df['LATITUDE'] = df_station_statistics.query("CD_ESTACAO == @station_code")['LATITUDE'].iloc[0]
    df['LONGITUDE'] = df_station_statistics.query("CD_ESTACAO == @station_code")['LONGITUDE'].iloc[0]
    df['ALTITUDE'] = df_station_statistics.query("CD_ESTACAO == @station_code")['ALTITUDE'].iloc[0]
    df_stations.append(df)
df_stations = pd.concat(df_stations)

df_stations = df_stations[['TOTAL_PRECIPITATION_mm', 'ATM_PRESSURE_mB', 'ATM_PRESSURE_SEA_LEVEL_mB', 
'MAX_ATM_PRESSURE_PREV_HOUR_mB', 'MIN_ATM_PRESSURE_PREV_HOUR_mB', 'GLOBAL_RADIATION_Kjm2', 
'AIR_TEMPERATURE_DRY_BULB_Celsius', 'DEW_POINT_TEMPERATURE_Celsius', 'MAX_TEMPERATURE_PREV_HOUR_Celsius', 
'MIN_TEMPERATURE_PREV_HOUR_Celsius', 'DEW_POINT_MAX_TEMPERATURE_PREV_HOUR_Celsius', 'DEW_POINT_MIN_TEMPERATURE_PREV_HOUR_Celsius',
'MAX_RELATIVE_HUMIDITY_PREV_HOUR_percentage', 'MIN_RELATIVE_HUMIDITY_PREV_HOUR_percentage', 'RELATIVE_HUMIDITY_percentage', 
'WIND_DIRECTION_degrees', 'WIND_MAX_GUNS_ms', 'WIND_SPEED_ms', 'WIND_SPEED_120m_ms', 
'LATITUDE', 'LONGITUDE', 'ALTITUDE', 'YEAR', 'MONTH', 'DAY', 'HOUR', 'DATETIME', 'NAME', 'CODE']]

df_stations.to_csv("./../../data/preprocessed/wind_data_conde_and_adjacent_stations.csv",index=False)

df_stations.groupby("YEAR").count()[['WIND_DIRECTION_degrees','WIND_SPEED_ms']]

df_stations.head(3)

df_stations.query("NAME == 'RIBEIRA DO AMPARO'").shape

df_stations.query("NAME == 'RIBEIRA DO AMPARO'").set_index('DATETIME').resample('1H').asfreq().shape
